third_party/blink/web_tests/wpt_internal/geolocation-api/not-enough-arguments.https.html

<!DOCTYPE html>
<html>
<head>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
</head>
<body>
<script type="module">
test(() => {
  const geolocation = window.navigator.geolocation;
  assert_throws_js(TypeError, () => geolocation.getCurrentPosition());
  assert_throws_js(TypeError, () => geolocation.watchPosition());
  assert_throws_js(TypeError, () => geolocation.clearWatch());
}, 'geolocation API rejects calls with missing arguments');
</script>
</body>
</html>
